+++ internal/bundler/bundler.go
@@ -1189,10 +1189,10 @@ func (s *scanner) maybeParseFile(
 
 	// Allow certain properties to be overridden
 	if len(resolveResult.JSXFactory) > 0 {
-		optionsClone.JSX.Factory = config.JSXExpr{Parts: resolveResult.JSXFactory}
+		optionsClone.JSX.Factory = config.DefineExpr{Parts: resolveResult.JSXFactory}
 	}
 	if len(resolveResult.JSXFragment) > 0 {
-		optionsClone.JSX.Fragment = config.JSXExpr{Parts: resolveResult.JSXFragment}
+		optionsClone.JSX.Fragment = config.DefineExpr{Parts: resolveResult.JSXFragment}
 	}
 	if resolveResult.UseDefineForClassFieldsTS != config.Unspecified {
 		optionsClone.UseDefineForClassFields = resolveResult.UseDefineForClassFieldsTS